feat: Alterações no frontend para a integração com o backend
