--- src/pages/candidatos/index.html
@@ -1,54 +1,0 @@
-<!doctype html>
-<html lang="pt-br">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Candidatos</title>
-  </head>
-
-  <body>
-    <h1>Candidatos</h1>
-    <form class="form">
-      <label for="name">Nome:</label><br>
-      <input name="name" type="text" id="name"><br>
-
-      <label for="birthDate">Data de Nascimento:</label><br>
-      <input name="birthDate" type="text" id="birthdate"><br>
-
-      <label for="course">Curso:</label><br>
-      <input name="course" type="text" id="course"><br>
-
-      <label for="period">Período:</label><br>
-      <input name="period" type="number" id="period"><br>
-
-      <label for="pronoun">Pronome:</label><br>
-      <select name="pronoum" id="pronoun">
-        <option value="">--Selecione uma opção--</option>
-        <option value="Ele/dele">Ele/dele</option>
-        <option value="Ela/dela">Ela/dela</option>
-        <option value="Elu/delu">Elu/delu</option>
-      </select><br>
-
-      <label for="gender">Sexo:</label><br>
-      <input name="gender" type="text" id="gender"><br>
-
-      <label for="motivation">Motivação:</label><br>
-      <input name="motivation" type="text" id="motivation"><br>
-
-      <label for="knowledgeLevel">Nível de Conhecimento:</label><br>
-      <input name="knowledgeLevel" type="text" id="knowledgeLevel"><br>
-
-      <label for="interestIDs">IDs de Interesse:</label><br>
-      <input name="interestIds" type="text" id="interestIds"><br>
-
-      <label for="email">E-Mail:</label><br>
-      <input name="email" type="text" id="email"><br>
-
-      <label for="password">Senha:</label><br>
-      <input name="password" type="text" id="password"><br>
-      
-      <button type="submit">Enviar</button>
-    </form>
-  </body>
-
-</html>
